mypy: type-clean the money modules + gate them in CI (ratchet)
Roadmap P1: mypy is configured but never run in CI. Fixed every type error in the
tractable money modules and added a CI gate so they stay clean.

Fixes (all behaviour-preserving — annotations/casts only):
- risk_engine: add a TYPE_CHECKING import for PortfolioTracker/MacroCalendar so the
  forward-ref annotations resolve. This cascaded: with self._portfolio properly
  typed, snapshot()→PortfolioState→equity_usd is float, which also cleared the two
  kelly no-any-return errors for free.
- portfolio: cast the pydantic model_copy() result (typed Any by the stubs) back to
  TradeExecution so _close_position_locked returns its declared type.
- trailing: annotate _STAGES as dict[int, dict[str, float | None]] so stage-def
  indexing is typed (was object).
- position_telemetry: float()-wrap the Wilder-ATR returns (list elements were Any).

CI now runs `mypy bot/risk bot/compliance bot/utils/trailing.py
bot/core/bitget_v3_client.py bot/core/position_telemetry.py` — the modules that are
fully clean today. live_executor (~31 errors) and the rest of bot/core are
deliberately NOT gated yet; they get added as they're cleaned, so the type floor
only ratchets up. Adds mypy==1.15.0 to the CI deps.

No runtime behaviour change. Full gate green; 285 risk/portfolio/trailing/telemetry
tests pass.

+++ bot/core/position_telemetry.py
@@ -93,11 +93,11 @@ def atr_from_candles(highs: list, lows: list, closes: list, period: int = 14) ->
     if not trs:
         return 0.0
     if len(trs) < period:
-        return sum(trs) / len(trs)
+        return float(sum(trs) / len(trs))
     atr = sum(trs[:period]) / period
     for tr in trs[period:]:
         atr = (atr * (period - 1) + tr) / period
-    return atr
+    return float(atr)
 
 
 def playbook_trail_threshold(direction: str, sl_trigger: float,
